+++ tests/p8_spec.lua
@@ -26,6 +26,16 @@ describe("p8", function()
       }
       assert.same(expected, p8.extract_strings_p8_content(p8_content))
     end)
+    it("should handle quotes in comments", function()
+      local code = [[
+pico-8 cartridge // http://www.pico-8.com
+version 43
+__lua__
+-- game's name
+-- by mike's friend
+]]
+      assert.same({}, p8.extract_strings_p8_content(code))
+    end)
   end)
   describe(".subst_l10n_strings()", function()
     it("should successfully substitute localized strings in .p8 file content", function()